+++ .gitignore
@@ -105,3 +105,4 @@ venv.bak/
 
 # idea 
 .vscode/
+.idea/
