+++ geoalarm/bot/constant.py
